feat: using `hive` database for save mod index
