+++ .vscode/settings.json
@@ -5,5 +5,8 @@
     "cSpell.language": "en,en-US",
     "yaml.schemas": {
         "https://json.schemastore.org/github-workflow.json": "file:///run/media/siongsng/Data/coding/Flutter/RPMLauncher/.github/workflows/build/linux.yml"
-    }
+    },
+    "cSpell.words": [
+        "murmurhash"
+    ]
 }
