+++ lib/launcher/Fabric/FabricClient.dart
@@ -5,7 +5,7 @@ import 'package:flutter/material.dart';
 import 'package:rpmlauncher/launcher/Fabric/FabricAPI.dart';
 import 'package:rpmlauncher/launcher/GameRepository.dart';
 import 'package:rpmlauncher/launcher/InstallingState.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/model/Game/MinecraftMeta.dart';
 import 'package:rpmlauncher/model/Game/MinecraftSide.dart';
 import 'package:rpmlauncher/model/IO/DownloadInfo.dart';
