+++ lib/launcher/Fabric/FabricServer.dart
@@ -14,7 +14,7 @@ import 'package:rpmlauncher/model/Game/MinecraftSide.dart';
 import 'package:path/path.dart';
 import 'package:rpmlauncher/launcher/Arguments.dart';
 import 'package:rpmlauncher/launcher/InstallingState.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/model/IO/DownloadInfo.dart';
 import 'package:rpmlauncher/util/I18n.dart';
 import 'package:rpmlauncher/util/util.dart';
