+++ lib/mod/CurseForge/Handler.dart
@@ -4,7 +4,7 @@ import 'dart:io';
 import 'package:dio/dio.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:rpmlauncher/launcher/APIs.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/util/I18n.dart';
 import 'package:rpmlauncher/util/RPMHttpClient.dart';
 import 'package:flutter/material.dart';
