+++ lib/model/Game/MinecraftVersion.dart
@@ -1,7 +1,7 @@
 import 'package:dio/dio.dart';
 import 'package:pub_semver/pub_semver.dart';
 import 'package:rpmlauncher/launcher/APIs.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/model/Game/MinecraftMeta.dart';
 import 'package:rpmlauncher/util/RPMHttpClient.dart';
 import 'package:rpmlauncher/util/util.dart';
