+++ lib/screen/Log.dart
@@ -21,7 +21,7 @@ import 'package:rpmlauncher/route/PushTransitions.dart';
 import 'package:rpmlauncher/screen/HomePage.dart';
 import 'package:rpmlauncher/util/Process.dart';
 import 'package:rpmlauncher/util/Config.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/util/I18n.dart';
 import 'package:rpmlauncher/util/util.dart';
 import 'package:rpmlauncher/util/data.dart';
