+++ lib/screen/VersionSelection.dart
@@ -6,7 +6,7 @@ import 'package:rpmlauncher/model/Game/MinecraftSide.dart';
 import 'package:rpmlauncher/model/Game/MinecraftVersion.dart';
 import 'package:rpmlauncher/screen/CurseForgeModPack.dart';
 import 'package:rpmlauncher/screen/FTBModPack.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/screen/RecommendedModpackScreen.dart';
 import 'package:rpmlauncher/util/I18n.dart';
 import 'package:flutter/material.dart';
