+++ lib/widget/AddInstance.dart
@@ -4,7 +4,7 @@ import 'package:rpmlauncher/launcher/Forge/ForgeClient.dart';
 import 'package:rpmlauncher/launcher/InstallingState.dart';
 import 'package:rpmlauncher/launcher/Vanilla/VanillaClient.dart';
 import 'package:rpmlauncher/launcher/Vanilla/VanillaServer.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
 import 'package:rpmlauncher/model/Game/Instance.dart';
 import 'package:rpmlauncher/model/Game/MinecraftMeta.dart';
 import 'package:rpmlauncher/model/Game/MinecraftSide.dart';
