+++ lib/widget/CurseForgeModVersion.dart
@@ -4,8 +4,8 @@ import 'dart:isolate';
 import 'package:flutter/material.dart';
 import 'package:path/path.dart';
 import 'package:rpmlauncher/mod/CurseForge/Handler.dart';
-import 'package:rpmlauncher/mod/ModLoader.dart';
-import 'package:rpmlauncher/model/Game/ModInfo.dart';
+import 'package:rpmlauncher/mod/mod_loader.dart';
+import 'package:rpmlauncher/model/Game/mod_info.dart';
 import 'package:rpmlauncher/model/IO/DownloadInfo.dart';
 import 'package:rpmlauncher/model/Game/Instance.dart';
 import 'package:rpmlauncher/model/IO/isolate_option.dart';
@@ -20,7 +20,7 @@ class CurseForgeModVersion extends StatefulWidget {
   final int curseID;
   final Directory modDir;
   final InstanceConfig instanceConfig;
-  final List<ModInfo> modInfos;
+  final Map<File, ModInfo> modInfos;
 
   const CurseForgeModVersion(
       {required this.curseID,
@@ -133,11 +133,12 @@ class _CurseForgeModVersionState extends State<CurseForgeModVersion> {
   }
 
   Future<Widget> installedWidget(Map fileInfo) async {
-    late ModInfo info;
+    late MapEntry<File, ModInfo> entry;
     try {
-      info = widget.modInfos
-          .firstWhere((info) => info.modHash == fileInfo["packageFingerprint"]);
-      installedFiles.add(info.file);
+      entry = widget.modInfos.entries.firstWhere(
+          (entry) => entry.value.murmur2Hash == fileInfo["packageFingerprint"]);
+
+      installedFiles.add(entry.key);
       return Column(
         mainAxisSize: MainAxisSize.min,
         children: [
