+++ pubspec.yaml
@@ -59,11 +59,14 @@ dependencies:
   window_manager: ^0.2.1
   file_picker: ^4.5.1
   rpmtw_dart_common_library: ^0.0.4
+  hive_flutter: ^1.1.0
 
 dev_dependencies:
+  build_runner: ^2.1.10
   flutter_lints: ^2.0.0
   flutter_test:
     sdk: flutter
+  hive_generator: ^1.1.2
   # msix: 2.3.0
 
 flutter:
